requirements.txt
flask
tensorflow
scikit-learn
pandas
numpy
yfinance
requests
python-dotenv
google-generativeai
finnhub-python